aula-4/server/requirements.txt
asgiref==3.4.1
bidict==0.21.2
bjoern==3.1.0
click==8.0.1
fastapi==0.66.0
fastavro==1.4.3
Flask==2.0.1
Flask-Cors==3.0.10
Flask-SocketIO==5.1.0
Flask-Sockets==0.2.1
gevent==21.1.2
gevent-websocket==0.10.1
greenlet==1.1.0
h11==0.12.0
itsdangerous==2.0.1
Jinja2==3.0.1
kafka-python==2.0.2
MarkupSafe==2.0.1
pydantic==1.8.2
python-dotenv==0.18.0
python-engineio==4.2.0
python-socketio==5.3.0
six==1.16.0
starlette==0.14.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
websockets==9.1
Werkzeug==2.0.1
zope.event==4.5.0
zope.interface==5.4.0
